code_files/freaking-math/requirements.txt
pywin32
keyboard
pyautogui
opencv-python
numpy
pytesseract
pillow